% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Queries the UK governments covid REST API and extracts case numbers by region.}
\usage{
get_data(
  filters,
  structure,
  endpoint = "https://api.coronavirus.data.gov.uk/v1/data"
)
}
\arguments{
\item{filters}{API filters. See the API documentations for
additional information: https://coronavirus.data.gov.uk/developers-guide}

\item{structure}{Structure parameter. See the API documentations
for additional information: https://coronavirus.data.gov.uk/developers-guide}

\item{endpoint}{Endpoint URL for the UK government data REST API.}
}
\value{
data.frame  Comprehensive dataframe containing all
                    the data for the given ``filter`` and ``structure`.`
}
\description{
Queries the UK governments covid REST API and extracts case numbers by region.
}
\examples{
query_filters <- c("areaType=ltla")
query_structure <- list(
  date       = "date",
  name       = "areaName",
  code       = "areaCode",
  daily      = "newCasesBySpecimenDate",
  cumulative = "cumCasesBySpecimenDate"
)
df_cases <- get_data(query_filters, query_structure)

query_filters <- c("areaType=nation", "areaName=england", "date=2020-04-01")
query_structure <- list(cumulative = "cumCasesBySpecimenDate")
df_cases <- get_data(query_filters, query_structure)
}
